grpc_server/requirements.txt
grpcio==1.62.1
protobuf==5.26.1
psycopg2-binary==2.9.9
